crest-annotation-cv/requirements.txt
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
black==23.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.4
cmake==3.25.0
coloredlogs==15.0.1
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.9.0
flatbuffers==24.3.25
fsspec==2025.2.0
h11==0.14.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
lit==15.0.7
MarkupSafe==2.1.2
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
opencv-python==4.11.0.86
packaging==23.1
pathspec==0.11.2
pillow==11.1.0
platformdirs==3.10.0
portalocker==3.1.1
protobuf==5.26.1
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.28.1
SAM-2 @ git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
sniffio==1.3.0
starlette==0.27.0
sympy==1.13.1
tomli==2.0.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==1.26.13
uvicorn==0.22.0
